moved function to dataAug
